% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_report.R
\name{eda_report}
\alias{eda_report}
\title{Exploratory Data Analysis Report}
\usage{
eda_report(
  X,
  M,
  A = NULL,
  Y = NULL,
  plot_corr = TRUE,
  plot_Y = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{X}{A data.frame or matrix containing predictors.}

\item{M}{A data.frame or matrix containing mediators.}

\item{A}{(optional) A vector of treatment assignments.}

\item{Y}{(optional) A vector of outcomes.}

\item{plot_corr}{Logical. If TRUE, a correlation matrix of X is plotted. Default is TRUE.}

\item{plot_Y}{Logical. If TRUE, density and box plots for Y are generated (if Y is provided). Default is TRUE.}

\item{verbose}{Logical. If TRUE, summary information is printed to the console. Default is TRUE.}
}
\value{
A list containing:
\describe{
  \item{summary}{A list with dimensions, missing values, and summary statistics for X, M, (and Y, A if provided).}
  \item{plots}{A list of ggplot objects, including the correlation matrix plot for X and density and box plots for Y (if provided).}
}
}
\description{
This function performs exploratory data analysis (EDA) for the input data.
It reports basic information such as dimensions, missing values, and summary statistics for predictors (X) and mediators (M).
If outcome (Y) is provided, density and box plots are generated, along with outlier information.
If treatment assignment (A) is provided, its distribution is also reported.
}
\examples{
\dontrun{
eda_results <- eda_report(X = sim_data$X,
                          M = sim_data$M,
                          A = sim_data$A,
                          Y = sim_data$Y,
                          plot_corr = TRUE,
                          plot_Y = TRUE,
                          verbose = TRUE)

# View EDA output
print(eda_results$summary)
# To view plots in an interactive environment, run:
print(eda_results$plots$scatter_plot_minvar)
print(eda_results$plots$scatter_plot_maxvar)
print(eda_results$plots$corr_plot)
print(eda_results$plots$Y_density)
print(eda_results$plots$Y_boxplot)
}

}
