% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_effects.R
\name{est_effects}
\alias{est_effects}
\title{Estimate Mediation Effects (ATE, NIE, NDE)}
\usage{
est_effects(
  X,
  Y,
  M,
  A,
  logY = FALSE,
  parallel = TRUE,
  nfold1 = 5,
  nfold2 = 3,
  net_alpha = 1,
  zeta = 0.1,
  optimizer = "pogs",
  intercept = FALSE,
  verbose = T
)
}
\arguments{
\item{X}{Design matrix (n x p) for continuous predictors.}

\item{Y}{Outcome vector (length n).}

\item{M}{Mediator matrix (n x q).}

\item{A}{Treatment assignment vector (binary, length n).}

\item{logY}{Logical. If TRUE, outcome Y is log-transformed. Default is FALSE.}

\item{parallel}{Logical. If TRUE, parallel computing is used. Default is TRUE.}

\item{nfold1}{Integer. Number of folds in cross-validation for outcome models. Default is 5.}

\item{nfold2}{Integer. Number of folds in cross-validation for mediator models. Default is 3.}

\item{net_alpha}{Numeric. Mixing parameter for elastic net in cv.glmnet (default is 1 for lasso).}

\item{zeta}{Numeric. Tuning parameter for residual balancing. Default is 0.1.}

\item{optimizer}{Character. Optimizer method for residual balancing. Default is "pogs".}

\item{intercept}{Logical. If TRUE, models include an intercept. Default is FALSE.}
}
\value{
A list containing:
\describe{
  \item{TE}{Total effect estimate.}
  \item{NIE}{Natural indirect effect estimate.}
  \item{NDE}{Natural direct effect estimate.}
  \item{Y11}{Debiased outcome estimate for the treatment group under treatment.}
  \item{Y10}{Debiased outcome estimate for the control group under the treatment counterfactual.}
  \item{Y00}{Debiased outcome estimate for the control group.}
}
}
\description{
This function estimates the total effect (TE), natural indirect effect (NIE), and natural direct effect (NDE)
using a debiased approach based on high-dimensional mediation analysis. It fits lasso models for the treatment
and control groups and then uses residual balancing for debiasing. Optionally, the outcome Y can be log-transformed.
}
\examples{
\dontrun{
  effects_result <- est_effects(X = sim_data$X,
                                Y = sim_data$Y,
                                M = sim_data$M,
                                A = sim_data$A,
                                 logY = FALSE,
                                 parallel = TRUE,
                                 nfold1 = 5,
                                 nfold2 = 5,
                                 net_alpha = 1,
                                 zeta = 0.5,
                                 optimizer = "pogs",
                                 intercept = FALSE,
                                 verbose = TRUE)
   
   print(effects_result)
}

}
